Fix bug with client-side for update_your_cp

+++ ShiftOS_TheReturn/ServerManager.cs
@@ -131,7 +131,7 @@ namespace ShiftOS.Engine
                     var args = JsonConvert.DeserializeObject<Dictionary<string, object>>(msg.Contents);
                     if(args["username"] as string == SaveSystem.CurrentSave.Username)
                     {
-                        SaveSystem.CurrentSave.Codepoints += (int)args["amount"];
+                        SaveSystem.CurrentSave.Codepoints += (long)args["amount"];
                         Infobox.Show($"MUD Control Centre", $"Someone bought an item in your shop, and they have paid {args["amount"]}, and as such, you have been granted these Codepoints.");
                         SaveSystem.SaveGame();
                     }
